feat(wave24): incremental integration — volatility_arb, trade_logger, momentum_backtest, tests
Extract submodules from wave24 branch into main:
- strategy/volatility_arb: split into analysis.rs + trade_lifecycle.rs submodules
- strategy/trade_logger: split into stats.rs, stats_rebuild.rs, summary.rs, write_flow.rs
- strategy/momentum_backtest: split into lifecycle.rs + results.rs submodules
- tests/: add native_async_traits, staging_workflow, workflow_migrations, workflow_security

All new modules compile cleanly with zero new errors introduced.
